+++ commitlint.config.js
@@ -1,0 +1,32 @@
+module.exports = {
+    'extends': ['@commitlint/config-conventional'],
+    'defaultIgnores': false,
+    'rules': {
+        'type-enum': [2, 'always', [
+            'feat',
+            'fix',
+            'add',
+            'test',
+            'refactor',
+            'perf',
+            'style',
+            'config',
+            'merge'
+        ]],
+        'scope-enum': [2, 'always', [
+            'api',
+            'factory',
+            'hotp',
+            'totp',
+            'generator',
+            'doc',
+            'deps',
+            'lint',
+            'branch',
+            'project'
+        ]],
+        'scope-empty': [2, 'never'],
+        'subject-min-length': [2, 'always', 5],
+        'subject-max-length': [2, 'always', 50],
+    }
+};