+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change Logs
 
+## v0.1.3
+
+- config(deps): replaced `tslint` with `eslint`.
+- config(deps): updated `@litert/encodings` to `v2.2.0`.
+
 ## v0.1.2
 
 - Fixed the `issuer` field.